Renamed package 'repository' to 'repositories' in Cosmic Metrics Collector
