+++ cosmic-metrics-collector/src/test/java/com/github/missioncriticalcloud/cosmic/metricscollector/repositories/VirtualMachineMetricsRepositoryIT.java
@@ -1,4 +1,4 @@
-package com.github.missioncriticalcloud.cosmic.metricscollector.repository;
+package com.github.missioncriticalcloud.cosmic.metricscollector.repositories;
 
 import static org.assertj.core.api.Assertions.assertThat;
